pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.2"
description = ""
authors = ["Pavel Ilchenko <arisesinmight@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pytest-cov = "^4.0.0"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
